Changes to Coatrack, DBdump, and fasta2DB.  Addition of DBa2b & DBb2a.  Internal improvements to DB library.
Catrack can now take multiple track names

DBa2b & DBb2a convert DBdump output to compressed binary and back

DBdump: the -p option is obsolete, -f to output a line a SMRT file
starts has been added.

fasta2DB: now also accepts ccs Pacbio data

DB.ch: Extensive changes so that now all data arrays, i.e. Reads,
Tracks, QVs, and Arrows can first be only “opened” meaning just the
read-index is loaded into memory.  Then individual items for a given
read can be loaded from the appropriate external file on demand, or all
the data/items can be loaded in toto.  That is, the interface for each
type is modeled on the interface originally developed for read data.

This allowed us to modify DBshow, DBdump, and DBstats so that they do
not load tracks in toto, giving a leaner memory footprint at no
significant decrease in speed.

Finally, the @-notation handling routines in DB.[ch] have been extended
to support @-signs in DB block references as well as
LAS block references.

+++ fasta2DAM.c
@@ -318,7 +318,7 @@ int main(int argc, char *argv[])
 
   { int            maxlen;
     int64          totlen, count[4];
-    int            rmax;
+    int64          rmax;
     DAZZ_READ      prec;
     char          *read;
     int            append;
@@ -467,7 +467,7 @@ int main(int argc, char *argv[])
                     break;
                   rlen += x;
                   if (rlen + MAX_NAME > rmax)
-                    { rmax = ((int) (1.2 * rmax)) + 1000 + MAX_NAME;
+                    { rmax = ((int64) (1.2 * rmax)) + 1000 + MAX_NAME;
                       read = (char *) realloc(read,rmax+1);
                       if (read == NULL)
                         { fprintf(stderr,"File %s.fasta, Line %d:",core,nline);
